# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Error {
  error: string;
  requestId?: string;
}

export interface User {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  role: string;
  createdAt: string;
}

/**
 * @nullable
 */
export type ProfileAddress = { [key: string]: unknown } | null;

export interface Profile {
  userId: string;
  /** @nullable */
  phone?: string | null;
  baseCurrency: string;
  /** @nullable */
  riskProfile?: string | null;
  /** @nullable */
  address?: ProfileAddress;
}

export interface KycStatus {
  id: string;
  userId: string;
  status: string;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  reviewedAt?: string | null;
  /** @nullable */
  reviewNote?: string | null;
}

export interface CurrentUser {
  user: User;
  profile: Profile;
  kyc: KycStatus;
}

export interface Portfolio {
  id: string;
  userId: string;
  name: string;
  baseCurrency: string;
  status: string;
  balance?: number;
}

export interface Holding {
  id: string;
  symbol: string;
  name: string;
  assetType: string;
  quantity: number;
  averageCost: number;
  currentPrice: number;
  marketValue?: number;
}

export interface Transaction {
  id: string;
  type: string;
  status: string;
  amount: number;
  currency: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface FundingRequest {
  id: string;
  type: string;
  status: string;
  amount: number;
  currency: string;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
}

export interface Notification {
  id: string;
  type: string;
  title: string;
  body: string;
  /** @nullable */
  readAt?: string | null;
  createdAt: string;
}

export interface Dashboard {
  user: User;
  profile?: Profile;
  kyc?: KycStatus;
  portfolio: Portfolio;
  holdings: Holding[];
  transactions: Transaction[];
  fundingRequests: FundingRequest[];
  notifications: Notification[];
  cashBalance?: number;
  portfolioValue?: number;
}

export type ProfileUpdateAddress = { [key: string]: unknown };

export interface ProfileUpdate {
  phone?: string;
  /**
     * @minLength 3
     * @maxLength 3
     */
  baseCurrency?: string;
  riskProfile?: string;
  address?: ProfileUpdateAddress;
}

export interface KycSubmission {
  /** @minLength 2 */
  fullName: string;
  dateOfBirth: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  country: string;
}

export interface FundingRequestInput {
  /** @exclusiveMinimum 0 */
  amount: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
}

export interface DepositIntent {
  requestId: string;
  clientSecret: string;
  /** @nullable */
  publishableKey?: string | null;
  status: string;
}

export interface FundingReview {
  status: string;
  reviewNote?: string;
}

/**
 * @nullable
 */
export type AuditLogMetadata = { [key: string]: unknown } | null;

export interface AuditLog {
  id: string;
  /** @nullable */
  actorUserId?: string | null;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  metadata?: AuditLogMetadata;
  createdAt: string;
}

export interface HealthStatus {
  status: string;
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Administrator access required
 */
export type ForbiddenResponse = Error;

/**
 * Request validation failed
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Capability is disabled by a release or compliance gate
 */
export type CapabilityDisabledResponse = Error;