# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kyc:
    get:
      operationId: getKycStatus
      tags: [auth]
      summary: Get current KYC status
      responses:
        "200":
          description: Current KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: submitKyc
      tags: [auth]
      summary: Submit KYC information for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmission"
      responses:
        "201":
          description: KYC submission accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /funding-requests:
    get:
      operationId: listFundingRequests
      tags: [funding]
      summary: List the authenticated user's funding requests
      responses:
        "200":
          description: Deposit and withdrawal requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FundingRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        requestId:
          type: string
...
    KycStatus:
      type: object
      required: [id, userId, status]
      properties:
        id: { type: string }
        userId: { type: string }
        status: { type: string }
        provider: { type: ["string", "null"] }
        submittedAt: { type: ["string", "null"], format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
        reviewNote: { type: ["string", "null"] }
    KycSubmission:
      type: object
      required: [fullName, dateOfBirth, country]
      properties:
        fullName: { type: string, minLength: 2 }
        dateOfBirth: { type: string, format: date }
        country: { type: string, minLength: 2, maxLength: 2 }
...
    FundingRequest:
      type: object
      required: [id, type, status, amount, currency, createdAt]
      properties:
        id: { type: string }
        type: { type: string }
        status: { type: string }
        amount: { type: number }
        currency: { type: string }
        stripePaymentIntentId: { type: ["string", "null"] }
        reviewNote: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...