# lib/api-zod/src/generated/types/fundingRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FundingRequest {
  id: string;
  type: string;
  status: string;
  amount: number;
  currency: string;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: Date;
}